General population utilities

General population utilities

GenPopUtil

Format

A data table (age, sex, GenPop_utility).

Source

Published parameter values describing health state utility values by age and sex by Ara and Brazier (2010) .

References

Ara R, Brazier JE (2010). “Populating an economic model with health state utility values: moving toward better practice.” Value in Health, 13(5), 509--518. https://doi.org/10.1111/j.1524-4733.2010.00700.x.