Adjust the condition-specific utility for age and sex

Calculates an average utility for each alcohol/tobacco-related condition by age and sex. Need to calculate the mean age and proportion of males for each condition in the HODAR data. Need to calculate the matched age-sex specific general population utility using the Ara and Brazier (2010) formula. Dividing these will give the multiplier by which to adjust the general population utility score by, to get age-adjusted condition-specific utilities.

Adj_cond_utils(
  Age_Sex_data,
  Condition_utilities,
  GenPopUtil = qalyr::GenPopUtil
)

Arguments

Age_Sex_data
mean age and proportion of males for in condition in data. Calculated using function MeanAge_Sex().

Condition_utilities
condition-specific utilities. Calculated using function CalcUtils().

GenPopUtil
table of general population utilities by age and sex. Default stored in package as qalyr::GenPopUtil.

Value

Returns a summary table of condition-specific utilities.

References

Ara R, Brazier JE (2010). “Populating an economic model with health state utility values: moving toward better practice.” Value in Health, 13(5), 509--518. doi: 10.1111/j.1524-4733.2010.00700.x , https://doi.org/10.1111/j.1524-4733.2010.00700.x.

Examples

if (FALSE) {

library(qalyr)

# Merge inpatient data and HODAR data and collapse to survey level data
data <- qalyr::merge_data(
  inpatient_data = qalyr::read_inpatient_data(),
  survey_data = qalyr::read_survey_data(),
  lkup = qalyr::lkup)

# Condition-specific utilities
cond_utils <- qalyr::CalcUtils(data, lkup = qalyr::lkup)

# Mean age and proportion of males
mean_age_sex <- qalyr::MeanAge_Sex(data, lkup = qalyr::lkup)

# age-sex-condition-specific utilities
adj_cond_utils <- qalyr::Adj_cond_utils(Age_Sex_data = mean_age_sex,
                                        Condition_utilities = cond_utils,
                                        GenPopUtil = qalyr::GenPopUtil)

}